Change the acceptance test nameespace
