+++ config/phpunit.xml
@@ -5,7 +5,7 @@
          xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/4.1/phpunit.xsd"
          backupGlobals="false"
          colors="true"
-         bootstrap="../var/bootstrap.php.cache"
+         bootstrap="../app/autoload.php"
 >
     <testsuites>
         <testsuite name="Project Test Suite">
@@ -21,7 +21,7 @@
 
     <filter>
         <whitelist>
-            <directory>../tests/PhpUnit</directory>
+            <directory>../tests</directory>
         </whitelist>
     </filter>
 </phpunit>
