+++ tests/Acceptance/Oath/TotpTest.php
@@ -1,11 +1,13 @@
 <?php
 
-namespace Tests\PhpUnit\Secrets;
+namespace Tests\Acceptance\Oath;
 
 use PHPUnit\Framework\TestCase;
 use GuzzleHttp\Client;
+use SURFnet\OATHBundle\OATH\TOTP;
+use SURFnet\OATHBundle\Services\Hash\Soft;
 
-class OcraTest extends TestCase
+class TotpTest extends TestCase
 {
     /**
      * @var Client
@@ -27,21 +29,6 @@ class OcraTest extends TestCase
     /**
      * @test
      */
-    public function getChallenge()
-    {
-        $response = $this->http->request('GET', 'oath/challenge/ocra', [
-            'headers' => [
-                'x-oathservice-consumerkey' => 'ThisKeyShouldBeSecret',
-            ]
-        ]);
-        $this->assertEquals(200, $response->getStatusCode());
-
-        $this->assertRegExp("/^\"[a-zA-Z0-9]{10}\"$/", $response->getBody()->getContents());
-    }
-
-    /**
-     * @test
-     */
     public function validateChallenge()
     {
         $response = $this->http->request('POST', 'secrets/id?secret=abcdef', [
@@ -51,7 +38,10 @@ class OcraTest extends TestCase
         ]);
         $this->assertEquals(204, $response->getStatusCode());
 
-        $response = $this->http->request('GET', 'oath/validate/ocra?challenge=ef46cb0560&response=682120&userId=id&sessionKey=3A4A', [
+        $totp = new TOTP(new Soft());
+        $totpResponse = $totp->calculateResponse('abcdef', 120, 6);
+
+        $response = $this->http->request('GET', 'oath/validate/totp?response='.$totpResponse.'&userId=id', [
             'headers' => [
                 'x-oathservice-consumerkey' => 'ThisKeyShouldBeSecret',
             ]
