+++ tests/Acceptance/Secrets/SecretsTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\PhpUnit\Secrets;
+namespace Tests\Acceptance\Secrets;
 
 use PHPUnit\Framework\TestCase;
 use GuzzleHttp\Client;
 
-class HotpTest extends TestCase
+class SecretsTest extends TestCase
 {
     /**
      * @var Client
@@ -27,22 +27,22 @@ class HotpTest extends TestCase
     /**
      * @test
      */
-    public function validateChallenge()
+    public function setSecret()
     {
-        $response = $this->http->request('POST', 'secrets/id?secret=abcdef', [
+        $response = $this->http->request('POST', 'secrets/id?secret=secret', [
             'headers' => [
                 'x-oathservice-consumerkey' => 'ThisKeyShouldBeSecret',
             ]
         ]);
-        $this->assertEquals(204, $response->getStatusCode());
 
-        $response = $this->http->request('GET', 'oath/validate/hotp?response=812453&userId=id', [
-            'headers' => [
-                'x-oathservice-consumerkey' => 'ThisKeyShouldBeSecret',
-            ]
-        ]);
         $this->assertEquals(204, $response->getStatusCode());
+    }
 
+    /**
+     * @test
+     */
+    public function deleteSecret()
+    {
         $response = $this->http->request('DELETE', 'secrets/id ', [
             'headers' => [
                 'x-oathservice-consumerkey' => 'ThisKeyShouldBeSecret',
